+++ include/screamer/rolling_mean.h
@@ -4,7 +4,7 @@
 #include <limits>
 #include <pybind11/pybind11.h>
 #include <pybind11/numpy.h>
-#include <screamer/buffer.h>
+#include "screamer/buffer.h"
 #include "screamer/base.h"
 
 namespace py = pybind11;